requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.14
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.34.2
gradio_client==1.10.3
greenlet==3.2.3
groovy==0.1.2
gTTS==2.5.4
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
multidict==6.5.1
mypy_extensions==1.1.0
narwhals==1.44.0
numpy==2.3.1
google-cloud-speech>=2.34.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
protobuf<6.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python_speech_features==0.6
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.1
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit==1.46.1
streamlit-audiorec==0.1.3
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
google-generativeai>=0.8.2
